Add files via upload
This notebook contains the codes used for my assessment in my MSc program. The task was to analyse census data of a particular region and recommend what could be built on an unoccupied piece of land, based on the analysis.
The code was written in python, and seaborn and pandas was used for analysis. I'm quite the learner, but this was fun.